refactor: Moved pagination to component
